+++ mjx/mujoco/mjx/_src/constraint.py
@@ -45,6 +45,7 @@ class _Efc(PyTreeNode):
   solref: jax.Array
   solimp: jax.Array
   margin: jax.Array
+  frictionloss: jax.Array
 
 
 def _kbi(
@@ -119,8 +120,9 @@ def _efc_equality_connect(m: Model, d: Data) -> Optional[_Efc]:
     j = (jacp1 - jacp2).T
     pos_imp = math.norm(pos)
     invweight = m.body_invweight0[obj1id, 0] + m.body_invweight0[obj2id, 0]
+    zero = jp.zeros_like(pos)
 
-    return _row(j, pos, pos_imp, invweight, solref, solimp, jp.zeros_like(pos))
+    return _row(j, pos, pos_imp, invweight, solref, solimp, zero, zero)
 
   args = (m.eq_obj1id, m.eq_obj2id, m.eq_data, m.eq_solref, m.eq_solimp)
   args = jax.tree_util.tree_map(lambda x: x[eq_id], args)
@@ -165,8 +167,9 @@ def _efc_equality_weld(m: Model, d: Data) -> Optional[_Efc]:
     pos_imp = math.norm(pos)
     invweight = m.body_invweight0[obj1id] + m.body_invweight0[obj2id]
     invweight = jp.repeat(invweight, 3, axis=0)
+    zero = jp.zeros_like(pos)
 
-    return _row(j, pos, pos_imp, invweight, solref, solimp, jp.zeros_like(pos))
+    return _row(j, pos, pos_imp, invweight, solref, solimp, zero, zero)
 
   args = (m.eq_obj1id, m.eq_obj2id, m.eq_data, m.eq_solref, m.eq_solimp)
   args = jax.tree_util.tree_map(lambda x: x[eq_id], args)
@@ -194,8 +197,9 @@ def _efc_equality_joint(m: Model, d: Data) -> Optional[_Efc]:
     j = jp.zeros((m.nv)).at[dofadr2].set(-deriv).at[dofadr1].set(1.0)
     invweight = m.dof_invweight0[dofadr1]
     invweight += m.dof_invweight0[dofadr2] * (obj2id > -1)
+    zero = jp.zeros_like(pos)
 
-    return _row(j, pos, pos, invweight, solref, solimp, jp.zeros_like(pos))
+    return _row(j, pos, pos, invweight, solref, solimp, zero, zero)
 
   args = (m.eq_obj1id, m.eq_obj2id, m.eq_data, m.eq_solref, m.eq_solimp)
   args = jax.tree_util.tree_map(lambda x: x[eq_id], args)
@@ -232,8 +236,9 @@ def _efc_equality_tendon(m: Model, d: Data) -> Optional[_Efc]:
     pos = pos1 - jp.dot(data[:5], dif_power)
     deriv = jp.dot(data[1:5], dif_power[:4] * jp.arange(1, 5)) * (obj2id > -1)
     j = jac1 + jac2 * -deriv
+    zero = jp.zeros_like(pos)
 
-    return _row(j, pos, pos, invweight, solref, solimp, jp.zeros_like(pos))
+    return _row(j, pos, pos, invweight, solref, solimp, zero, zero)
 
   inv1, inv2 = m.tendon_invweight0[obj1id], m.tendon_invweight0[obj2id]
   jac1, jac2 = d.ten_J[obj1id], d.ten_J[obj2id]
@@ -245,9 +250,32 @@ def _efc_equality_tendon(m: Model, d: Data) -> Optional[_Efc]:
 
 
 def _efc_friction(m: Model, d: Data) -> Optional[_Efc]:
-  # TODO(robotics-team): implement _instantiate_friction
-  del m, d
-  return None
+  """Calculates constraint rows for dof frictionloss."""
+  dof_id = np.nonzero(m.dof_hasfrictionloss)[0]
+  tendon_id = np.nonzero(m.tendon_hasfrictionloss)[0]
+
+  size = dof_id.size + tendon_id.size
+  if (m.opt.disableflags & DisableBit.FRICTIONLOSS) or (size == 0):
+    return None
+
+  args_dof = (jp.eye(m.nv), m.dof_frictionloss, m.dof_invweight0, m.dof_solref)
+  args_dof += (m.dof_solimp,)
+  args_dof = jax.tree_util.tree_map(lambda x: x[dof_id], args_dof)
+
+  args_ten = (d.ten_J, m.tendon_frictionloss, m.tendon_invweight0)
+  args_ten += (m.tendon_solref_fri, m.tendon_solimp_fri)
+  args_ten = jax.tree_util.tree_map(lambda x: x[tendon_id], args_ten)
+
+  args = jax.tree_util.tree_map(
+      lambda *x: jp.concatenate(x), args_dof, args_ten
+  )
+
+  @jax.vmap
+  def rows(j, frictionloss, invweight, solref, solimp):
+    z = jp.zeros_like(frictionloss)
+    return _row(j, z, z, invweight, solref, solimp, z, frictionloss)
+
+  return rows(*args)
 
 
 def _efc_limit_ball(m: Model, d: Data) -> Optional[_Efc]:
@@ -267,9 +295,10 @@ def _efc_limit_ball(m: Model, d: Data) -> Optional[_Efc]:
     active = pos < 0
     j = jp.zeros(m.nv).at[jp.arange(3) + dofadr].set(-axis)
     invweight = m.dof_invweight0[dofadr]
+    z = jp.zeros_like(pos)
 
     return _row(
-        j * active, pos * active, pos, invweight, solref, solimp, jnt_margin
+        j * active, pos * active, pos, invweight, solref, solimp, jnt_margin, z
     )
 
   args = (m.jnt_qposadr, m.jnt_dofadr, m.jnt_range, m.jnt_margin, m.jnt_solref)
@@ -296,9 +325,10 @@ def _efc_limit_slide_hinge(m: Model, d: Data) -> Optional[_Efc]:
     active = pos < 0
     j = jp.zeros(m.nv).at[dofadr].set((dist_min < dist_max) * 2 - 1)
     invweight = m.dof_invweight0[dofadr]
+    z = jp.zeros_like(pos)
 
     return _row(
-        j * active, pos * active, pos, invweight, solref, solimp, jnt_margin
+        j * active, pos * active, pos, invweight, solref, solimp, jnt_margin, z
     )
 
   args = (m.jnt_qposadr, m.jnt_dofadr, m.jnt_range, m.jnt_margin, m.jnt_solref)
@@ -332,8 +362,11 @@ def _efc_limit_tendon(m: Model, d: Data) -> Optional[_Efc]:
   pos = jp.minimum(dist_min, dist_max) - margin
   active = pos < 0
   j = jax.vmap(jp.multiply)(j, ((dist_min < dist_max) * 2 - 1) * active)
+  zero = jp.zeros_like(pos)
 
-  return jax.vmap(_row)(j, pos * active, pos, invweight, solref, solimp, margin)
+  return jax.vmap(_row)(
+      j, pos * active, pos, invweight, solref, solimp, margin, zero
+  )
 
 
 def _efc_contact_frictionless(m: Model, d: Data) -> Optional[_Efc]:
@@ -362,6 +395,7 @@ def _efc_contact_frictionless(m: Model, d: Data) -> Optional[_Efc]:
         c.solref,
         c.solimp,
         c.includemargin,
+        jp.zeros_like(pos),
     )
 
   contact = jax.tree_util.tree_map(lambda x: x[con_id], d.contact)
@@ -406,6 +440,7 @@ def _efc_contact_pyramidal(m: Model, d: Data, condim: int) -> Optional[_Efc]:
         c.solref,
         c.solimp,
         c.includemargin,
+        jp.zeros_like(pos),
     )
 
   contact = jax.tree_util.tree_map(lambda x: x[con_id], d.contact)
@@ -450,6 +485,7 @@ def _efc_contact_elliptic(m: Model, d: Data, condim: int) -> Optional[_Efc]:
         solref,
         c.solimp,
         c.includemargin,
+        jp.zeros_like(pos),
     )
 
   contact = jax.tree_util.tree_map(lambda x: x[con_id], d.contact)
@@ -460,7 +496,8 @@ def _efc_contact_elliptic(m: Model, d: Data, condim: int) -> Optional[_Efc]:
 def counts(efc_type: np.ndarray) -> Tuple[int, int, int, int]:
   """Returns equality, friction, limit, and contact constraint counts."""
   ne = (efc_type == ConstraintType.EQUALITY).sum()
-  nf = 0  # no support for friction loss yet
+  nf = (efc_type == ConstraintType.FRICTION_DOF).sum()
+  nf += (efc_type == ConstraintType.FRICTION_TENDON).sum()
   nl = (efc_type == ConstraintType.LIMIT_JOINT).sum()
   nl += (efc_type == ConstraintType.LIMIT_TENDON).sum()
   nc_f = (efc_type == ConstraintType.CONTACT_FRICTIONLESS).sum()
@@ -487,6 +524,20 @@ def make_efc_type(
     num_rows += (m.eq_type == EqType.JOINT).sum()
     num_rows += (m.eq_type == EqType.TENDON).sum()
     efc_types += [ConstraintType.EQUALITY] * num_rows
+
+  if not m.opt.disableflags & DisableBit.FRICTIONLOSS:
+    nf_dof = (
+        m.dof_hasfrictionloss.sum()
+        if isinstance(m, Model)
+        else (m.dof_frictionloss > 0).sum()
+    )
+    efc_types += [ConstraintType.FRICTION_DOF] * nf_dof
+    nf_tendon = (
+        m.tendon_hasfrictionloss.sum()
+        if isinstance(m, Model)
+        else (m.tendon_frictionloss > 0).sum()
+    )
+    efc_types += [ConstraintType.FRICTION_TENDON] * nf_tendon
 
   if not m.opt.disableflags & DisableBit.LIMIT:
     efc_types += [ConstraintType.LIMIT_JOINT] * m.jnt_limited.sum()
@@ -570,12 +621,12 @@ def make_constraint(m: Model, d: Data) -> Data:
     k, b, imp = _kbi(m, efc.solref, efc.solimp, efc.pos_imp)
     r = jp.maximum(efc.invweight * (1 - imp) / imp, mujoco.mjMINVAL)
     aref = -b * (efc.J @ d.qvel) - k * imp * efc.pos_aref
-    return aref, r, efc.pos_aref + efc.margin, efc.margin
+    return aref, r, efc.pos_aref + efc.margin, efc.margin, efc.frictionloss
 
-  aref, r, pos, margin = fn(efc)
+  aref, r, pos, margin, frictionloss = fn(efc)
   d = d.replace(
       efc_J=efc.J, efc_D=1 / r, efc_aref=aref, efc_pos=pos, efc_margin=margin
   )
-  d = d.replace(efc_frictionloss=jp.zeros_like(r))
+  d = d.replace(efc_frictionloss=frictionloss)
 
   return d